// src/app/_dto/TblHMenu.dto.ts
import { TbldMenuCoDto } from "./TbldMenuCoDto.dto";


export class TblHMenu {

    HMenu_SysID: number;
    HMenu_Code: string;
    HMenu_Name: string;
    HMenu_AcDe_Yn: string;
    HMenu_AcDe_SysID: number;
    HMenu_AcDe_Code: string;
    HMenu_AcDe_Name: string;

    TbldMenuCo :TbldMenuCoDto[] = []

}